refactor: move Record into request package
