+++ request/safeslice.go
@@ -1,4 +1,4 @@
-package record
+package request
 
 import "sync"
